+++ Processes/performance-forecasting/pf-service/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'java'
 }
 
-group = 'com.bluemarlin'
+group = 'org.apache.bluemarlin'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '1.8'
