+++ Processes/performance-forecasting/pf-service/src/main/java/org/apache/bluemarlin/pfservice/model/Day.java
@@ -16,7 +16,7 @@
  *  limitations under the License.
  */
 
-package com.bluemarlin.pfservice.model;
+package org.apache.bluemarlin.pfservice.model;
 
 import java.util.*;
