+++ Processes/performance-forecasting/pf-service/src/main/java/org/apache/bluemarlin/pfservice/util/BMLogger.java
@@ -16,7 +16,7 @@
  *  limitations under the License.
  */
 
-package com.bluemarlin.pfservice.util;
+package org.apache.bluemarlin.pfservice.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
